+++ src/main/java/com/github/chainmailstudios/astromine/client/screen/FluidMixerHandledScreen.java
@@ -27,7 +27,7 @@ public class FluidMixerHandledScreen extends DefaultedEnergyFluidHandledScreen<F
 
 		WHorizontalArrow arrow = mainPanel.createChild(WHorizontalArrow::new, Position.of(secondInputFluidBar, secondInputFluidBar.getWidth() + 9, secondInputFluidBar.getHeight() / 2 - 8, 8), Size.of(22, 16))
 				.setLimitSupplier(() -> mixer.limit)
-				.setProgressSupplier(() -> mixer.current);
+				.setProgressSupplier(() -> (int) mixer.current);
 
 		WFluidVolumeFractionalVerticalBar outputFluidBar = mainPanel.createChild(WFluidVolumeFractionalVerticalBar::new, Position.of(arrow, arrow.getWidth() + 7, -secondInputFluidBar.getHeight() / 2 + 8, 8), Size.of(fluidBar));
